+++ spec/lib/00_sqlite_aware_spec.rb
@@ -9,6 +9,7 @@ require_relative 'filter_model/affix_synonyms_spec'
 require_relative 'dynamic_listener_spec'
 require_relative 'filterable_block_spec'
 require_relative 'auto_detect_backend_shared_examples'
+require_relative 'filter_model/range_filters_shared_examples'
 
 require 'support/database_manager'
 
@@ -27,6 +28,8 @@ RSpec.describe "SQLite" do
   include_examples "FilterModel::FilterMapBlockDSL", db
   include_examples "FilterModel::AffixSynonyms", db
   include_examples "AutoDetectBackend", db
+  include_examples "FilterModel::RangeFilters", db
+  include_examples "FilterModel::NestedRangeFilters", db
   # Also run the Filterable block DSL shared examples (DB-agnostic)
   include_examples "Filterable::FilterMapBlockDSL"
 end